+++ config.js
@@ -1,5 +1,6 @@
 'use strict';
 /* eslint no-process-env: 0 */
+const AWS = require('aws-sdk');
 
 const env = process.env.NODE_ENV;
 const useMocks = process.env.USE_MOCKS === 'true' || !env;
@@ -26,11 +27,10 @@ module.exports = {
     secret: process.env.KEYCLOAK_SECRET
   },
   awsSqs: {
-    region: process.env.AWS_REGION || 'eu-west-2',
+    endpoint: new AWS.Endpoint('http://localhost:4566'),
+    region: process.env.AWS_REGION,
     queueUrl: process.env.SQS_URL,
-    credentials: {
-      accessKeyId: process.env.ACCESS_KEY_ID,
-      secretAccessKey: process.env.SECRET_ACCESS_KEY
-    }
+    accessKeyId: process.env.ACCESS_KEY_ID,
+    secretAccessKey: process.env.SECRET_ACCESS_KEY
   }
 };
